refactor: refactoring class file path and revise api spec
