+++ src/main/java/com/lotdiz/notificationservice/controller/restcontroller/NotificationRestController.java
@@ -14,9 +14,11 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
+@RequestMapping("/api")
 @RequiredArgsConstructor
 public class NotificationRestController {
   private final UnreadNotificationService unreadNotificationService;
@@ -45,7 +47,7 @@ public class NotificationRestController {
                 .build());
   }
 
-  @GetMapping("notifications/unread-count")
+  @GetMapping("/notifications/unread-count")
   public ResponseEntity<SuccessResponse<Map<String, Long>>> numberOfUnreadNotifications() {
     Long memberId = 1L;
     Long numberOfUnreadNotifications =
